requirements.txt
dj-database-url==0.4.1
django-analytical==2.0.0
django-postgres-copy==0.0.7
djangorestframework==3.3.2
gunicorn==19.4.5
Markdown==2.6.2
MySQL-python==1.2.5
pbr==1.8.0
psycopg2==2.6.1
python-http-client==1.2.3
requests==2.9.1
sendgrid==2.2.1
six==1.9.0
smtpapi==0.3.1
stevedore==1.8.0
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wheel==0.24.0
whitenoise==3.0
